Move document service middleware to separate file
